#27 Rename module (extras/replacement → extras/replacing)
